+++ src/app/router.tsx
@@ -23,8 +23,8 @@ import { mapInfoFilter } from "../api/services/map/types";
 import { vendorItemFilter } from "../api/services/vendor/types";
 import { skillFilter } from "../api/services/skill/types";
 import { Redirect } from "../lib/tsr/react/Redirect";
-import { RouterLocation } from "../lib/tsr/Route";
 import { zodLiteralString } from "../lib/zod/zodLiteralString";
+import { RouterLocation } from "../lib/tsr/types";
 import { requireAuth } from "./util/requireAuth";
 import { requireSettings } from "./util/requireSettings";
 import { t } from "./tsr";
