+++ src/lib/tsr/utils/normalizeLocation.ts
@@ -1,5 +1,5 @@
 import { Location } from "history";
-import { RouterLocation } from "../Route";
+import { RouterLocation } from "../types";
 
 export function normalizeLocation(location: Location): RouterLocation {
   return (location.pathname + location.search) as RouterLocation;
